Capacity sensors integration.
This is limited functionality version.
Not supporting single sensor configuration on chanels.
This limitation is caused by the use of two different
capacitive sensors. One sensor is NO - Normal Open,
the second one is NC - Normal Closed. Sensor state
detection is basing on meassuring voltage on the
sensor output. The NC type has max voltage value
while, the NO has 0 voltage. This cause that driver
program "0" state is representing sum of voltage from
both sensors, and it is max voltage. In case of two
capacitive sensor of the same type the problem does
exist - the program "0" state is representing 0 voltage
in case of NO type sensor, for example:

Program state    NO_1    NO_2   Out
0                0V      0V     0V
1                5V      0V     5V
2                0V      7V     7V
3                5V      7V     12V

Therefore there is no issue with detection single or
double sensors on the port input.

Must be updated when appripriate sensors will be applied.